Bones, Camera, Vertex Groups, Update
Added a bunch of changes all at once because I've been tweaking this on
and off for a couple of years with no version control.

Added in bone/helper to armature import. Currently loses data to
distinguish bones from helpers, needs to be addressed.

Added in support for single camera import, very hardcoded, sans
rotation adjustments. Uses an 'empty' object to do camera position
targeting.

Added in Vertex Groups and associations to bones as a TON of Hook
modifiers. This was the sanest way I could find of doing this in
blender since the way vertex grouping is done is rather different,
limiting vertices to a single group. Please correct if wrong.

Added a bunch of mesh import changes and updates to make it work with a
more recent version of blender.